Restructured schema stuff to remove the chained dependency
