+++ src/open_targets_platform_mcp/tools/__init__.py
@@ -7,16 +7,11 @@ from open_targets_platform_mcp.tools.batch_query.batch_query import (
 )
 from open_targets_platform_mcp.tools.query.query import query_with_jq, query_without_jq
 from open_targets_platform_mcp.tools.schema.schema import (
-    get_open_targets_graphql_schema,
-    prefetch_schema,
-)
-from open_targets_platform_mcp.tools.schema.subschema import (
     get_categories_for_docstring,
-    prefetch_category_subschemas,
+    get_open_targets_graphql_schema,
 )
 from open_targets_platform_mcp.tools.schema.type_graph import (
     get_type_dependencies,
-    prefetch_type_graph,
 )
 from open_targets_platform_mcp.tools.search_entities.search_entities import (
     search_entities,
@@ -28,9 +23,6 @@ __all__ = [
     "get_categories_for_docstring",
     "get_open_targets_graphql_schema",
     "get_type_dependencies",
-    "prefetch_category_subschemas",
-    "prefetch_schema",
-    "prefetch_type_graph",
     "query_with_jq",
     "query_without_jq",
     "search_entities",
