+++ src/open_targets_platform_mcp/tools/schema/helper/__init__.py
@@ -1,0 +1,27 @@
+"""Helper package for Schema Tools."""
+
+from open_targets_platform_mcp.tools.schema.helper.graph import (
+    TypeGraph,
+    build_type_graph,
+    get_reachable_types,
+    get_reachable_types_with_depth,
+)
+from open_targets_platform_mcp.tools.schema.helper.subschema import (
+    CategorySubschema,
+    CategorySubschemas,
+)
+from open_targets_platform_mcp.tools.schema.helper.utils import (
+    load_categories,
+    types_to_sdl,
+)
+
+__all__ = [
+    "CategorySubschema",
+    "CategorySubschemas",
+    "TypeGraph",
+    "build_type_graph",
+    "get_reachable_types",
+    "get_reachable_types_with_depth",
+    "load_categories",
+    "types_to_sdl",
+]
